+++ CHANGELOG.md
@@ -10,6 +10,11 @@ This file contains the notable changes in the ISLa project since version 0.2a1
 - Added a method `ISLaSolver.fuzz()` that produces one solution at each call, i.e., not a generator of
   solutions as returned by `ISLaSolver.solver()`.
 
+### Changed
+
+- Better support for unsatisfiable formulas: States with unsatisfiable SMT-LIB atoms in their formulas
+  are discarded early. Before, it could happen that the solver diverged.
+
 ## [0.8.16] - 2022-08-17
 
 ### Added